Fix signed/unsigned warning in filter-thunder.c

+++ source/fitz/filter-thunder.c
@@ -25,8 +25,8 @@ next_thunder(fz_context *ctx, fz_stream *stm, size_t max)
 	unsigned char *ep;
 	int c, v, i, pixels, index;
 
-	if (max > state->len)
-		max = state->len;
+	if (max > (size_t)state->len)
+		max = (size_t)state->len;
 
 	ep = p + max;
